+++ CHANGELOG.md
@@ -25,7 +25,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
--
+- `Trainer.add_argparse_args` classmethod fixed. Now it adds a type for the arguments.
 
 ## [0.7.1] - 2020-03-07